Takeover IOptions
