+++ src/Microsoft.Extensions.Options/IConfigureOptions.cs
@@ -12,7 +12,8 @@ namespace Microsoft.Extensions.Options
         /// <summary>
         /// Invoked to configure a TOptions instance.
         /// </summary>
+        /// <param name="name">The name of the options instance being configured.</param>
         /// <param name="options">The options instance to configure.</param>
-        void Configure(TOptions options);
+        void Configure(string name, TOptions options);
     }
 }
