+++ test/Microsoft.Extensions.Options.Test/Microsoft.Extensions.Options.Test.csproj
@@ -19,4 +19,8 @@
     <PackageReference Include="xunit" Version="$(XunitVersion)" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}" />
+  </ItemGroup>
+
 </Project>
